Use a new version file.
